Update Card component
