+++ content/README.md
@@ -86,7 +86,7 @@ title: "Example Post" # required
 slug: "/example" # optional: if not specified, will be generated by the file's relative path
 no_background: false # optional: if true, will not contain content with the rounded gray panel background
 ---
-import EventCards from "../components/EventCards"
+import Card from "../components/Card"
 <!-- Import path is relative to the MDX file -->
 
 # Header
@@ -94,7 +94,7 @@ import EventCards from "../components/EventCards"
 Here are some awesome events:
 
 <!-- You can use React components! -->
-<EventCards includeSeries={["uiuctf"]} />
+<Card heading="Heading" title="Title" ... />
 ```
 
 ## *Profile*
