+++ src/styles/main.css
@@ -128,24 +128,6 @@ a:hover {
   background-color: var(--color-secondary);
   color: var(--color-background);
 }
-.card {
-  background-color: var(--color-surface);
-  border-color: var(--color-surface);
-  border-radius: 0.75rem;
-  isolation: isolate;
-  overflow: hidden;
-  position: relative;
-  box-shadow: none;
-  bottom: 0;
-  right: 0;
-  line-height: 1.2;
-  transition: 0.1s;
-}
-.card-hover:hover .card {
-  box-shadow: 14px 14px var(--color-primary);
-  bottom: 7px;
-  right: 7px;
-}
 /* fancy navbar toggle on mobile */
 .icon-bar {
 	width: 22px;
